Move to k6/secrets and add example
