+++ examples/secrets/file.secret
@@ -1,0 +1,2 @@
+cool=some
+else=source
