+++ internal/js/modules/k6/secrets/secrets_test.go
@@ -1,0 +1,122 @@
+package secrets
+
+import (
+	"testing"
+
+	"github.com/stretchr/testify/assert"
+	"github.com/stretchr/testify/require"
+
+	"go.k6.io/k6/internal/secretsource/mock"
+	"go.k6.io/k6/js/modulestest"
+	"go.k6.io/k6/secretsource"
+)
+
+func testRuntimeWithSecrets(t testing.TB, secretSources map[string]secretsource.SecretSource) *modulestest.Runtime {
+	testRuntime := modulestest.NewRuntime(t)
+	var err error
+	testRuntime.VU.InitEnvField.SecretsManager, _, err = secretsource.NewSecretsManager(secretSources)
+	require.NoError(t, err)
+
+	m, ok := New().NewModuleInstance(testRuntime.VU).(*Secrets)
+	require.True(t, ok)
+	require.NoError(t, testRuntime.VU.RuntimeField.Set("secrets", m.Exports().Default))
+
+	return testRuntime
+}
+
+func TestSecrets(t *testing.T) {
+	t.Parallel()
+
+	type secretsTest struct {
+		secretsources map[string]secretsource.SecretSource
+		script        string
+		expectedValue any
+		expectedError string
+	}
+
+	cases := map[string]secretsTest{
+		"simple": {
+			secretsources: map[string]secretsource.SecretSource{
+				"default": mock.NewMockSecretSource("some", map[string]string{
+					"secret": "value",
+				}),
+			},
+			script:        "secrets.get('secret')",
+			expectedValue: "value",
+		},
+		"error": {
+			secretsources: map[string]secretsource.SecretSource{
+				"default": mock.NewMockSecretSource("some", map[string]string{
+					"secret": "value",
+				}),
+			},
+			script:        "secrets.get('not_secret')",
+			expectedError: "no value",
+		},
+		"multiple": {
+			secretsources: map[string]secretsource.SecretSource{
+				"default": mock.NewMockSecretSource("some", map[string]string{
+					"secret": "value",
+				}),
+				"second": mock.NewMockSecretSource("some", map[string]string{
+					"secret2": "value2",
+				}),
+			},
+			script:        "secrets.get('secret')",
+			expectedValue: "value",
+		},
+		"multiple get default": {
+			secretsources: map[string]secretsource.SecretSource{
+				"default": mock.NewMockSecretSource("some", map[string]string{
+					"secret": "value",
+				}),
+				"second": mock.NewMockSecretSource("some", map[string]string{
+					"secret2": "value2",
+				}),
+			},
+			script:        "secrets.source('default').get('secret')",
+			expectedValue: "value",
+		},
+		"multiple get not default": {
+			secretsources: map[string]secretsource.SecretSource{
+				"default": mock.NewMockSecretSource("some", map[string]string{
+					"secret": "value",
+				}),
+				"second": mock.NewMockSecretSource("some", map[string]string{
+					"secret2": "value2",
+				}),
+			},
+			script:        "secrets.source('second').get('secret2')",
+			expectedValue: "value2",
+		},
+		"get secret without source": {
+			secretsources: map[string]secretsource.SecretSource{},
+			script:        "secrets.get('secret')",
+			expectedError: "no source with name default",
+		},
+		"get none existing source": {
+			secretsources: map[string]secretsource.SecretSource{
+				"default": mock.NewMockSecretSource("some", map[string]string{
+					"secret": "value",
+				}),
+			},
+			script:        "secrets.source('second') != undefined",
+			expectedValue: true,
+		},
+	}
+
+	for name, testCase := range cases {
+		t.Run(name, func(t *testing.T) {
+			t.Parallel()
+			testruntime := testRuntimeWithSecrets(t, testCase.secretsources)
+
+			v, err := testruntime.RunOnEventLoop(testCase.script)
+			if testCase.expectedError != "" {
+				require.ErrorContains(t, err, testCase.expectedError)
+				return
+			}
+			require.NoError(t, err)
+			assert.Equal(t, testCase.expectedValue, v.Export())
+		})
+	}
+}
