+++ js/modules/k6/k6.go
@@ -13,7 +13,6 @@ import (
 	"go.k6.io/k6/js/modules"
 	"go.k6.io/k6/lib"
 	"go.k6.io/k6/metrics"
-	"go.k6.io/k6/secretsource"
 )
 
 var (
@@ -31,8 +30,7 @@ type (
 
 	// K6 represents an instance of the k6 module.
 	K6 struct {
-		vu             modules.VU
-		secretsManager *secretsource.SecretsManager
+		vu modules.VU
 	}
 )
 
@@ -49,15 +47,11 @@ func New() *RootModule {
 // NewModuleInstance implements the modules.Module interface to return
 // a new instance for each VU.
 func (*RootModule) NewModuleInstance(vu modules.VU) modules.Instance {
-	return &K6{vu: vu, secretsManager: vu.InitEnv().SecretsManager}
+	return &K6{vu: vu}
 }
 
 // Exports returns the exports of the k6 module.
 func (mi *K6) Exports() modules.Exports {
-	s, err := mi.secrets()
-	if err != nil {
-		common.Throw(mi.vu.Runtime(), err)
-	}
 	return modules.Exports{
 		Named: map[string]interface{}{
 			"check":      mi.Check,
@@ -65,38 +59,8 @@ func (mi *K6) Exports() modules.Exports {
 			"group":      mi.Group,
 			"randomSeed": mi.RandomSeed,
 			"sleep":      mi.Sleep,
-			"secrets":    s,
 		},
 	}
-}
-
-func (mi *K6) secrets() (*sobek.Object, error) {
-	obj, err := secretSourceObjectForSourceName(mi.vu.Runtime(), mi.secretsManager, secretsource.DefaultSourceName)
-	if err != nil {
-		return nil, err
-	}
-
-	err = obj.Set("source", func(sourceName string) (*sobek.Object, error) {
-		return secretSourceObjectForSourceName(mi.vu.Runtime(), mi.secretsManager, sourceName)
-	})
-	if err != nil {
-		return nil, err
-	}
-
-	return obj, nil
-}
-
-func secretSourceObjectForSourceName(
-	rt *sobek.Runtime, manager *secretsource.SecretsManager, sourceName string,
-) (*sobek.Object, error) {
-	obj := rt.NewObject()
-	err := obj.Set("get", func(key string) (string, error) {
-		return manager.Get(sourceName, key)
-	})
-	if err != nil {
-		return nil, err
-	}
-	return obj, nil
 }
 
 // Fail is a fancy way of saying `throw "something"`.
